requirements.txt
Babel==2.3.4
ciso8601==1.0.2
debtcollector==1.6.0
enum34==1.1.6
eventlet==0.19.0
fasteners==0.14.1
funcsigs==1.0.2
greenlet==0.4.10
iso8601==0.1.11
Jinja2==2.8
influxdb==2.8.0
keystoneauth1==2.9.0
MarkupSafe==0.23
monotonic==1.1
msgpack-python==0.4.7
netaddr==0.7.18
netifaces==0.10.4
oslo.concurrency==3.12.0
oslo.config==3.13.0
oslo.context==2.6.0
oslo.i18n==3.8.0
oslo.log==3.12.0
oslo.middleware==3.15.0
oslo.serialization==2.11.0
oslo.service==1.13.0
oslo.utils==3.15.0
Paste==2.0.3
PasteDeploy==1.5.2
pbr==1.10.0
positional==1.1.1
prettytable==0.7.2
pyinotify==0.9.6
python-dateutil==2.5.3
python-keystoneclient==3.2.0
python-monascaclient==1.2.0
pytz==2016.6.1
PyYAML==3.11
repoze.lru==0.6
requests==2.10.0
retrying==1.3.3
rfc3986==0.3.1
Routes==2.3.1
six==1.10.0
stevedore==1.16.0
WebOb==1.6.1
wheel==0.24.0
wrapt==1.10.8
yos-social-sdk==0.0.2
